move SemanticHash to API package
